memorized-question-answering-system 0.5.0
updates:
* Renaming the Service
