+++ pom.xml
@@ -4,9 +4,9 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>eu.wse-research</groupId>
-    <artifactId>fake-question-answering-system</artifactId>
-    <version>0.4.4</version>
-    <name>Fake Question Answering System</name>
+    <artifactId>memorized-question-answering-system</artifactId>
+    <version>0.5.0</version>
+    <name>MemorizedQuestionAnsweringSystem</name>
 
     <parent>
         <groupId>org.springframework.boot</groupId>
@@ -22,7 +22,7 @@
         <log4j.version>2.20.0</log4j.version>
         <google.guava.version>[32.0.0,)</google.guava.version>
         <docker.image.prefix>wseresearch</docker.image.prefix>
-        <docker.image.name>fake-question-answering-system</docker.image.name>
+        <docker.image.name>memorized-question-answering-system</docker.image.name>
         <dockerfile-maven-version>1.4.10</dockerfile-maven-version>
     </properties>
 
@@ -153,7 +153,7 @@
                     </execution>
                 </executions>
                 <configuration>
-                    <mainClass>eu.wseresearch.fakequestionansweringsystem.Application</mainClass>
+                    <mainClass>eu.wseresearch.memorizedquestionansweringsystem.Application</mainClass>
                 </configuration>
             </plugin>
             <!-- https://github.com/spotify/dockerfile-maven -->
