+++ src/main/resources/config/application.properties
@@ -1,6 +1,6 @@
 # update the port number
 server.port=40100
-spring.application.name=FakeQuestionAnsweringSystem
+spring.application.name=MemorizedQuestionAnsweringSystem
 
 # define what additional (local) properties files should be used
 spring.profiles.active=local
@@ -10,7 +10,7 @@ spring.output.ansi.enabled=always
 
 # log level definitions
 # change logging level in production
-logging.file.name=FakeQuestionAnsweringSystem.log
+logging.file.name=MemQASystem.log
 logging.file.path=/var/log/
 logging.level.*=INFO
 logging.level.org.springframework.core=WARN
@@ -22,7 +22,7 @@ logging.level.org.apache.catalina.core=WARN
 logging.level.org.springframework.jmx=WARN
 logging.level.org.springframework.mock=WARN
 logging.level.org.springframework.test=WARN
-logging.level.eu.wseresearch.fakequestionansweringsystem=INFO
+logging.level.eu.wseresearch.memorizedquestionansweringsystem=INFO
 
 # springdoc options
 springdoc.version=1.6.11
@@ -30,7 +30,7 @@ springdoc.swagger-ui.path=/swagger-ui.html
 springdoc.api-docs.path=/api-docs
 
 # caching
-fakequestionansweringsystem.sparql.cache.specs=expireAfterAccess=86400s
+memqasystem.sparql.cache.specs=expireAfterAccess=86400s
 
 # cross-origin
 cross.origin=*
@@ -52,3 +52,4 @@ qado.resultset.maxrows=10000
 #virtuoso.graph=
 #virtuoso.username=
 #virtuoso.password=
+#virtuoso.query.timeout=
